More progress
- Removed extraneous call to set GR_PYTHON_DIR
- Fixed up edacs3.grc
- Removed problematic callbacks
- Minor cleanup in find_chan_nums_impl
- Refactored proc_msg_impl to be more c++ like
- Fixed issue with float in trunked_radio.py

+++ python/trunked_radio.py
@@ -67,7 +67,7 @@ class trunked_radio(gr.hier_block2):
         _fp_sel_index_thread.start()
             
         self.rational_resampler_xxx_1 = filter.rational_resampler_fff(
-                interpolation=sink_rate / 8000,
+                interpolation=int(sink_rate / 8000),
                 decimation=1,
                 taps=None,
                 fractional_bw=None,
